refactor: split the workspace settings page into sections
