+++ eslint.config.js
@@ -262,7 +262,6 @@ export default defineConfigWithVueTs(
             'resources/js/pages/channels/Show.vue',
             'resources/js/pages/teams/Analytics.vue',
             'resources/js/pages/teams/AuditExports.vue',
-            'resources/js/pages/teams/Edit.vue',
             'resources/js/pages/teams/Groups.vue',
             'resources/js/pages/teams/integrations/Bot.vue',
             'resources/js/pages/teams/integrations/Index.vue',
